refactor rules and spacy model
