+++ Pipfile
@@ -31,6 +31,7 @@ black = "*"
 opencensus-ext-azure = "*"
 pytest-asyncio = "*"
 pytest-snapshot = "*"
+eng = "*"
 
 [requires]
 python_version = "3.9"
