+++ app/models.py
@@ -1,4 +1,3 @@
-from numpy import array
 from pydantic import BaseModel, validator
 from typing import Dict, List
 from typing import Optional
@@ -10,9 +9,10 @@ import string
 
 class Lang(object):
     def __init__(self, locale):
-        self.locale = locale[0:2]
+        self.lang = locale[0:2]
+        self.locale = locale
 
-        if self.locale == "en":
+        if self.lang == "en":
             trans_locale = "en_GB"
         else:
             trans_locale = "de_DE"
@@ -216,21 +216,12 @@ class ResultOut(BaseModel):
 
         rewrite_to_swiss_german = False
         if (
-            lang.locale == "de"
+            lang.lang == "de"
             and "de-CH" in config.preferred_variants
             and config.preferred_variants.find("de-CH")
             <= config.preferred_variants.find("de")
         ):
             rewrite_to_swiss_german = True
-
-        rewrite_to_british = False
-        if (
-            lang.locale == "en"
-            and "en-GB" in config.preferred_variants
-            and config.preferred_variants.find("en-GB")
-            <= config.preferred_variants.find("en")
-        ):
-            rewrite_to_british = True
 
         cleaned_alternatives = []
         for i, alternative in enumerate(alternatives):
@@ -239,9 +230,6 @@ class ResultOut(BaseModel):
 
             if rewrite_to_swiss_german:
                 alternative = alternative.replace("ß", "ss")
-
-            if rewrite_to_british:
-                alternative = alternative.replace("color", "colour")
 
             if "~" in alternative:
                 variants = alternative.split("~")
@@ -321,7 +309,7 @@ class ResultsOut(BaseModel):
     language: str
 
     def factory(results, lang):
-        return ResultsOut(results, lang.locale)
+        return ResultsOut(results, lang.lang)
 
     factory = staticmethod(factory)
