+++ app/rules.py
@@ -1,0 +1,88 @@
+import pandas as pd
+import copy
+
+rules = {
+    "de-DE": {
+        # load agentic language
+        "df_agentic_ct": "agentic.csv",
+        # load Gender denom
+        "df_gender_ct": "gendered_noun.csv",
+        # load Gender denom false_positives
+        "genderdenom_false_positives": "titles_false_positives.csv",
+        # load style words
+        "df_style_word": "style_words.csv",
+        "df_style_sentences": "style_sentences.csv",
+        # load openly discriminating words de
+        "df_open_dis_word": "open_dis_words.csv",
+        "df_open_dis_sentence": "open_dis_sentences.csv",
+        # load unconscious_bias word (nouns, not nouns) and sentences de
+        "df_ub_noun_word": "ub_noun_words.csv",
+        "df_ub_no_noun_word": "ub_no_noun_words.csv",
+        "df_ub_sentences": "ub_sentences.csv",
+        # load inslusive words
+        "df_d_and_i_words": "d_and_i_words.csv",
+        # load inslusive sentences
+        "df_d_and_i_words_sentences": "d_and_i_sentences.csv",
+        # load communal coded terms
+        "df_communal_words": "communal.csv",
+    },
+    "en-US": {
+        # load agentic language
+        "df_agentic_words": "agentic.csv",
+        # load openly discriminating words
+        "df_open_dis_word": "open_dis_words.csv",
+        "df_open_dis_sentence": "open_dis_sentences.csv",
+        # load inclusive language
+        "df_inclusive_word": "inclusive_words.csv",
+        "df_inclusive_sentence": "inclusive_sentences.csv",
+        # load style words
+        "df_style_word": "style_words.csv",
+        "df_style_sentence": "style_sentences.csv",
+        # load gendered language
+        "df_gendered_no_noun_word": "gendered_no_noun_words.csv",
+        "df_gendered_sentence": "gendered_sentences.csv",
+        "df_gendered_noun_word": "gendered_noun_words.csv",
+    },
+}
+
+rules["en-GB"] = rules["en-US"].copy()
+
+for locale in rules:
+    for csv in rules[locale]:
+        rules[locale][csv] = pd.read_csv(
+            "training_data/" + locale + "/" + rules[locale][csv]
+        )
+
+# list of "hollow word" sentences
+rules["de-DE"]["terms_style"] = list(rules["de-DE"]["df_style_sentences"]["Lemma"])
+
+# list of "d_and_i_words word" sentences
+rules["de-DE"]["terms_d_and_i_words"] = list(
+    rules["de-DE"]["df_d_and_i_words_sentences"]["Lemma"]
+)
+
+# dictionaries to handle false positives
+rules["de-DE"]["false_positive_agentic_const"] = [
+    "selbst",
+    "flexible",
+    "Probleme",
+    "unabhängig",
+    "Entwickler",
+]
+rules["de-DE"]["false_positive_style"] = ["international"]
+rules["de-DE"]["exceptions"] = [
+    "Unternehmen",
+    "Firma",
+    "Gruppe",
+    "Gesellschaft",
+    "Kollektivgesellschaft",
+    "Team",
+    "Organization",
+    "Gliederung",
+]
+rules["de-DE"]["gender_false_positive"] = rules["de-DE"]["genderdenom_false_positives"][
+    "False_positives"
+].tolist()
+
+rules["de-CH"] = copy.deepcopy(rules["de-DE"])
+rules["de-AT"] = copy.deepcopy(rules["de-DE"])
