+++ training_data/en-US/gendered_no_noun_words.csv
@@ -1,0 +1,18 @@
+Lemma,Word_Type,Priority,Category,Primary_subcategory,Sg_all_split,Pl_all_split,Alt_split
+attention-seeking,a,1,gendered,female_stereotype,"['outgoing', 'at ease in social situations', 'gregarious', 'lively', 'extroverted', 'unrestrained', 'validation-seeking', 'approval-seeking', 'causing disruption', 'tending to cause disruption']",[''],"['outgoing', 'at ease in social situations', 'gregarious', 'lively', 'extroverted', 'unrestrained', 'validation-seeking', 'approval-seeking', 'causing disruption', 'tending to cause disruption']"
+attentionseeking,a,1,gendered,female_stereotype,"['outgoing', 'at ease in social situations', 'gregarious', 'lively', 'extroverted', 'unrestrained', 'validation-seeking', 'approval-seeking', 'causing disruption', 'tending to cause disruption']",[''],"['outgoing', 'at ease in social situations', 'gregarious', 'lively', 'extroverted', 'unrestrained', 'validation-seeking', 'approval-seeking', 'causing disruption', 'tending to cause disruption']"
+beautiful,a,,gendered,female_stereotype,"['attractive', 'good-looking']",[''],"['attractive', 'good-looking']"
+bossy,a,,gendered,female_stereotype,"['strict', 'in charge', 'supervising', 'responsible for outcomes/results', 'direct', 'communicating expectations']",[''],"['strict', 'in charge', 'supervising', 'responsible for outcomes/results', 'direct', 'communicating expectations']"
+handsome,a,,gendered,male_stereotype,"['attractive', 'good-looking']",[''],"['attractive', 'good-looking']"
+housewife,a,,gendered,female_stereotype,"['homemaker', 'they']",[''],"['homemaker', 'they']"
+man,v,,gendered,hidden_image,"['staff', 'crew', 'occupy', 'work', 'use']",[''],"['staff', 'crew', 'occupy', 'work', 'use']"
+man-made,a,,gendered,hidden_image,"['synthetic', 'machine-made', 'artificial']",[''],"['synthetic', 'machine-made', 'artificial']"
+man-sized,a,,gendered,hidden_image,"['very large', 'generous', 'calling for the strength traditionally attributed to a man']",[''],"['very large', 'generous', 'calling for the strength traditionally attributed to a man']"
+manmade,a,,gendered,hidden_image,"['synthetic', 'machine-made', 'artificial']",[''],"['synthetic', 'machine-made', 'artificial']"
+mansized,a,,gendered,hidden_image,"['very large', 'generous', 'calling for the strength traditionally attributed to a man']",[''],"['very large', 'generous', 'calling for the strength traditionally attributed to a man']"
+manwards,adv,,gendered,hidden_image,"['of humankind', 'at humanity', 'toward humankind']",[''],"['of humankind', 'at humanity', 'toward humankind']"
+manwise,adv,,gendered,hidden_image,"['in the manner of a human being', 'like a human being']",[''],"['in the manner of a human being', 'like a human being']"
+outflank,v,,gendered,male_stereotype,"['outwit', 'gain a tactical advantage over', 'win an advantage over']",[''],"['outwit', 'gain a tactical advantage over', 'win an advantage over']"
+outrank,v,,gendered,male_stereotype,"['be more important', 'improve upon', 'hold a higher position', 'have a higher rank']",[''],"['be more important', 'improve upon', 'hold a higher position', 'have a higher rank']"
+submissive,a,1,gendered,female_stereotype,"['cooperative', 'open to compromise', 'keen to keep peace', 'willing to accommodate (others)']",[''],"['cooperative', 'open to compromise', 'keen to keep peace', 'willing to accommodate (others)']"
+whine,v,1,gendered,female_stereotype,"['complain', 'grouse', 'grumble', 'express doubts', 'be open about their doubts', 'annoy by repeatedly complaining', 'annoy (someone) with repeated complaints']",[''],"['complain', 'grouse', 'grumble', 'express doubts', 'be open about their doubts', 'annoy by repeatedly complaining', 'annoy (someone) with repeated complaints']"
