+++ training_data/en-US/inclusive_sentences.csv
@@ -1,0 +1,13 @@
+Lemma,Priority,Category,Primary_subcategory
+affinity group,,inclusive,d_and_i
+different background,1,inclusive,d_and_i
+do one's share,1,inclusive,communal
+gender identity,,inclusive,d_and_i
+implicit bias,,inclusive,d_and_i
+impostor syndrome,,inclusive,d_and_i
+of all ages,1,inclusive,d_and_i
+of all beliefs,1,inclusive,d_and_i
+of all religions,1,inclusive,d_and_i
+of all sexual orientations,1,inclusive,d_and_i
+part time,1,inclusive,communal
+sexual orientation,1,inclusive,d_and_i
