+++ training_data/en-US/inclusive_words.csv
@@ -1,0 +1,157 @@
+Lemma,Priority,Category,Primary_subcategory
+accessibility,,inclusive,d_and_i
+affectionate,1,inclusive,communal
+ages,1,inclusive,d_and_i
+agree,1,inclusive,communal
+agreeable,1,inclusive,communal
+ally,,inclusive,d_and_i
+amplification,1,inclusive,d_and_i
+belief,1,inclusive,d_and_i
+beliefs,1,inclusive,d_and_i
+belong,1,inclusive,d_and_i
+care,1,inclusive,communal
+caring,1,inclusive,communal
+cheer,1,inclusive,communal
+cheerful,1,inclusive,communal
+child,1,inclusive,communal
+children,1,inclusive,communal
+cis-gender,,inclusive,d_and_i
+cis-gendered,,inclusive,d_and_i
+code-switching,,inclusive,d_and_i
+collaborate,1,inclusive,communal
+collaboration,1,inclusive,communal
+collaborative,1,inclusive,communal
+commit,1,inclusive,communal
+communal,1,inclusive,communal
+communicate,1,inclusive,communal
+communication,1,inclusive,communal
+communicative,1,inclusive,communal
+community,1,inclusive,communal
+compassion,1,inclusive,communal
+compassionate,1,inclusive,communal
+connect,1,inclusive,communal
+connection,1,inclusive,communal
+considerate,1,inclusive,communal
+cooperate,1,inclusive,communal
+cooperation,1,inclusive,communal
+cooperative,1,inclusive,communal
+create,1,inclusive,communal
+creative,1,inclusive,communal
+creativity,1,inclusive,communal
+culture,1,inclusive,d_and_i
+curious,1,inclusive,communal
+depend,1,inclusive,communal
+dependable,1,inclusive,communal
+dependence,1,inclusive,communal
+dependent,1,inclusive,communal
+discrimination,1,inclusive,d_and_i
+discuss,1,inclusive,communal
+discussion,1,inclusive,communal
+diverse,1,inclusive,d_and_i
+diversify,1,inclusive,d_and_i
+diversity,,inclusive,d_and_i
+emotion,1,inclusive,communal
+emotional,1,inclusive,communal
+empathetic,1,inclusive,communal
+empathic,,inclusive,communal
+empathize,1,inclusive,communal
+empathy,1,inclusive,communal
+engage,1,inclusive,communal
+engagement,1,inclusive,communal
+enthuse,1,inclusive,communal
+enthusiasm,1,inclusive,communal
+enthusiast,1,inclusive,communal
+enthusiastic,1,inclusive,communal
+equality,1,inclusive,d_and_i
+inclusify,1,inclusive,d_and_i
+equity,,inclusive,d_and_i
+fair,1,inclusive,d_and_i
+faithful,,inclusive,communal
+feel,1,inclusive,communal
+flatter,,inclusive,communal
+flattering,1,inclusive,communal
+friendly,1,inclusive,d_and_i
+fulfilling,1,inclusive,d_and_i
+gentle,1,inclusive,communal
+grow,1,inclusive,communal
+happy,1,inclusive,d_and_i
+help,1,inclusive,communal
+helpful,,inclusive,communal
+honest,1,inclusive,communal
+honesty,1,inclusive,communal
+identity,1,inclusive,d_and_i
+impact,1,inclusive,communal
+impactful,1,inclusive,communal
+inclusion,1,inclusive,d_and_i
+inclusive,1,inclusive,d_and_i
+initiative,1,inclusive,communal
+integrity,1,inclusive,d_and_i
+inter-dependence,,inclusive,communal
+inter-dependent,1,inclusive,communal
+inter-personal,1,inclusive,communal
+interest,1,inclusive,communal
+intersectionality,,inclusive,d_and_i
+job-sharing,1,inclusive,communal
+kind,1,inclusive,communal
+kinship,1,inclusive,communal
+learn,1,inclusive,communal
+learner,1,inclusive,communal
+loyal,1,inclusive,communal
+loyalty,1,inclusive,communal
+maintain,1,inclusive,communal
+meaningful,1,inclusive,d_and_i
+mediate,1,inclusive,communal
+mediation,1,inclusive,communal
+mentor,1,inclusive,communal
+mentoring,1,inclusive,communal
+modest,1,inclusive,communal
+modesty,1,inclusive,communal
+moral,,inclusive,communal
+motivate,1,inclusive,communal
+motivation,1,inclusive,communal
+need,1,inclusive,communal
+nurture,1,inclusive,communal
+participate,1,inclusive,communal
+pleasant,1,inclusive,communal
+pleased,1,inclusive,communal
+polite,1,inclusive,communal
+precise,1,inclusive,communal
+precision,1,inclusive,communal
+quiet,1,inclusive,communal
+relaxed,1,inclusive,d_and_i
+reliable,1,inclusive,communal
+religion,1,inclusive,d_and_i
+request,1,inclusive,communal
+resilience,1,inclusive,communal
+resilient,1,inclusive,communal
+respect,1,inclusive,d_and_i
+respectful,1,inclusive,d_and_i
+responding,1,inclusive,communal
+responsibility,1,inclusive,communal
+responsible,1,inclusive,communal
+responsive,1,inclusive,communal
+safe,1,inclusive,d_and_i
+sensitive,1,inclusive,communal
+sensitivity,1,inclusive,communal
+share,1,inclusive,communal
+sharing,,inclusive,communal
+steadfast,,inclusive,communal
+support,1,inclusive,communal
+supportive,1,inclusive,communal
+sustainable,1,inclusive,communal
+sympathetic,1,inclusive,communal
+sympathize,,inclusive,communal
+sympathy,1,inclusive,communal
+tender,1,inclusive,communal
+thoughtful,1,inclusive,communal
+together,1,inclusive,communal
+trust,1,inclusive,communal
+trustworthy,1,inclusive,communal
+under-represented,1,inclusive,d_and_i
+understand,1,inclusive,communal
+understanding,,inclusive,communal
+understood,1,inclusive,communal
+warm,1,inclusive,communal
+warmth,1,inclusive,communal
+woman,1,inclusive,communal
+yield,1,inclusive,communal
