+++ training_data/en-US/open_dis_sentences.csv
@@ -1,0 +1,80 @@
+Lemma,Priority,Category,Primary_subcategory,Alt_Field,Alt_split
+African queen,,openly_discriminating,homophobia,-,['-']
+anal assassin,,openly_discriminating,homophobia,-,['-']
+anal astronaut,,openly_discriminating,homophobia,-,['-']
+arse bandit,,openly_discriminating,homophobia,-,['-']
+ass bandit,,openly_discriminating,homophobia,-,['-']
+Aunt Jemima,,openly_discriminating,racism,-,['-']
+baby mama,,openly_discriminating,sexism,-,['-']
+bathroom predator,,openly_discriminating,homophobia,-,['-']
+batty boy,,openly_discriminating,homophobia,-,['-']
+bean queen,,openly_discriminating,homophobia,-,['-']
+Beastie Boy,,openly_discriminating,antisemitism,-,['-']
+boat niggers,,openly_discriminating,racism,-,['-']
+boat people,,openly_discriminating,xenophobia,-,['-']
+bone smuggler,,openly_discriminating,homophobia,-,['-']
+booty buffer,,openly_discriminating,homophobia,-,['-']
+border jumper,,openly_discriminating,xenophobia,-,['-']
+brown piper,,openly_discriminating,homophobia,-,['-']
+brownie king,,openly_discriminating,homophobia,-,['-']
+brownie queen,,openly_discriminating,homophobia,-,['-']
+burkha bitch,,openly_discriminating,antimuslim,-,['-']
+camel jockey,,openly_discriminating,xenophobia,-,['-']
+cardboard nigger,,openly_discriminating,racism,-,['-']
+carpet mincher,,openly_discriminating,homophobia,-,['-']
+chicken queen,,openly_discriminating,homophobia,-,['-']
+cock tease,,openly_discriminating,sexism,-,['-']
+cockpipe cosmonaut,,openly_discriminating,homophobia,-,['-']
+cockstruction worker,,openly_discriminating,homophobia,-,['-']
+cotton picker,,openly_discriminating,racism,-,['-']
+dark meat,,openly_discriminating,racism,-,['-']
+diaper head,,openly_discriminating,antimuslim,-,['-']
+Diesel dyke,,openly_discriminating,homophobia,-,['-']
+dog eater,,openly_discriminating,xenophobia,-,['-']
+dog muncher,,openly_discriminating,xenophobia,-,['-']
+dot head,,openly_discriminating,xenophobia,-,['-']
+dot or feather?,,openly_discriminating,racism,-,['-']
+drag dyke,,openly_discriminating,homophobia,-,['-']
+dune coon,,openly_discriminating,antimuslim,-,['-']
+fag bomb,,openly_discriminating,homophobia,-,['-']
+fag hag,,openly_discriminating,sexism,-,['-']
+feather head,,openly_discriminating,racism,-,['-']
+fender head,,openly_discriminating,xenophobia,-,['-']
+fresh off the boat,,openly_discriminating,xenophobia,-,['-']
+friend of Dorothy,,openly_discriminating,homophobia,-,['-']
+frog eater,,openly_discriminating,xenophobia,-,['-']
+frog nibbler,,openly_discriminating,xenophobia,-,['-']
+gig nose,,openly_discriminating,antisemitism,-,['-']
+Gillette blade,,openly_discriminating,homophobia,-,['-']
+girlie man,,openly_discriminating,homophobia,-,['-']
+goat fucker,,openly_discriminating,antimuslim,-,['-']
+grey queen,,openly_discriminating,homophobia,-,['-']
+Hop Sing,,openly_discriminating,racism,-,['-']
+house boy,,openly_discriminating,racism,-,['-']
+jew boy,,openly_discriminating,antisemitism,-,['-']
+Jew Yorker,,openly_discriminating,antisemitism,-,['-']
+Jewish prince,,openly_discriminating,antisemitism,-,['-']
+Jewish princess,,openly_discriminating,antisemitism,-,['-']
+jungle fever,,openly_discriminating,racism,-,['-']
+lady boy,,openly_discriminating,transphobia,-,['-']
+light in the fedora,,openly_discriminating,homophobia,-,['-']
+light in the loafers,,openly_discriminating,homophobia,-,['-']
+light in the pants,,openly_discriminating,homophobia,-,['-']
+limp wrist,,openly_discriminating,homophobia,-,['-']
+meat masseuse,,openly_discriminating,homophobia,-,['-']
+muscle Mary,,openly_discriminating,homophobia,-,['-']
+oven dodger,,openly_discriminating,antisemitism,-,['-']
+pancake face,,openly_discriminating,racism,-,['-']
+potato queen,,openly_discriminating,homophobia,-,['-']
+rice queen,,openly_discriminating,homophobia,-,['-']
+ring raider,,openly_discriminating,homophobia,-,['-']
+Salsa queen,,openly_discriminating,homophobia,-,['-']
+sand nigger,,openly_discriminating,antimuslim,-,['-']
+Suzy Homemaker,,openly_discriminating,sexism,-,['-']
+taco queen,,openly_discriminating,homophobia,-,['-']
+trophy wife,,openly_discriminating,sexism,-,['-']
+Uncle Tom,,openly_discriminating,racism,-,['-']
+welfare queen,,openly_discriminating,sexism,-,['-']
+white trash,,openly_discriminating,racism,-,['-']
+window licker,1,openly_discriminating,ableism,-,['-']
+woolie woofter,,openly_discriminating,homophobia,-,['-']
