+++ training_data/en-US/style_sentences.csv
@@ -1,0 +1,19 @@
+Lemma,Priority,Category,Primary_subcategory,Alt_split
+amongst others,1,style,filler,['-']
+career opportunity,1,style,hollow,['']
+contemporary working conditions,1,style,hollow,['']
+dream job,1,style,hollow,['']
+dream tasks,1,style,hollow,['']
+dynamic team,1,style,hollow,['']
+flat hierarchy,1,style,hollow,['']
+flat organisation,1,style,hollow,['']
+flat organization,1,style,hollow,['']
+high quality,1,style,hollow,['']
+highly motivated,1,style,hollow,['']
+most of all,1,style,hollow,['']
+organisational talent,1,style,hollow,['']
+outside the box,1,style,hollow,['']
+team player,1,style,hollow,"['person who values team effort over individual gains', 'person who enjoys collaboration', 'someone who can compromise to reach a shared goal']"
+team spirit,1,style,hollow,['a mindset that values team effort over individual gains']
+think outside of the box,1,style,hollow,['']
+training budget,1,style,hollow,['']
