+++ training_data/en-US/style_words.csv
@@ -1,0 +1,17 @@
+Lemma,Priority,Category,Primary_subcategory,Alt_split
+etc.,1,style,filler,['-']
+flexible,1,style,hollow,['']
+global,1,style,hollow,['']
+innovation,1,style,hollow,['']
+innovative,1,style,hollow,['']
+normal,1,style,hollow,['typically']
+passion,1,style,hollow,"['enthusiasm', 'dedication', 'committment', 'fascination', 'keen interest', 'love (for something)', 'excitement (about doing something)']"
+passionate,1,style,hollow,"['enthusiastic', 'dedicated', 'committed (to something)', 'fascinated with', 'keenly interested in', 'excited (about doing something)']"
+quick,1,style,hollow,['']
+really,1,style,filler,['-']
+state-of-the-art,1,style,hollow,['']
+team-player,1,style,hollow,"['person who values team effort over individual gains', 'person who enjoys collaboration', 'someone who can compromise to reach a shared goal']"
+teamplayer,1,style,hollow,"['person who values team effort over individual gains', 'person who enjoys collaboration', 'someone who can compromise to reach a shared goal']"
+unique,1,style,hollow,['']
+workforce,1,style,hollow,"['employees', 'personnel']"
+worldwide,1,style,hollow,"['in (...) countries', '(...) offices across (...) continents']"
